move more code to helpers
